Define a variable with the initial time
To assign the official shift time, we must to refer to the same scale of time as the other data. That's why we assign this value in the variable "initial.time" before and we won't need to make conditional if the first data recorded have been from stage record or gaugings.