Manage adjustement.type to nearest flood to MAP but without stage record
In fact, if the user chose to assign shift time to nearest flood peak to MAP, there is not mandatory to have stage record for all gauging data. That's why the code bug if there isn't flood peak to associate. In this case, BayDERS will take the MAP criterion because it is the only way to treat